+++ R/ciftiTools-package.R
@@ -8,6 +8,7 @@
 #'  \item{\code{as.xifti}:}{   Combine numeric data to form a \code{"xifti"}}
 #'  \item{\code{read_surf}:}{   Read in a surface GIFTI file as a \code{"surf"}}
 #'  \item{\code{info_cifti}:}{   Read the metadata in a CIFTI file}
+#'  \item{\code{load_surf}:}{   Read in a surface geometry included in \code{ciftiTools}}
 #'  \item{\code{load_parc}:}{   Read in a dlabel parcellation included in \code{ciftiTools}}
 #' }
 #'