+++ R/data.R
@@ -1,0 +1,6 @@
+#' Files included in \code{ciftiTools}
+#'
+#' @name ciftiTools.files
+#' @docType data
+#' @keywords data
+NULL