+++ R/make_xifti.R
@@ -183,7 +183,7 @@ make_xifti <- function(
     if (HCP_32k_auto_mwall & nrow(xifti$data$cortex_left) == 29696) {
       if (is.null(xifti$meta$cortex$medial_wall_mask$left)) {
         # from ciftiTools R/sysdata.rda
-        xifti$meta$cortex$medial_wall_mask$left <- HCP_32k_mwall_template[,1]
+        xifti$meta$cortex$medial_wall_mask$left <- ciftiTools.data$HCP_32k_mwall_template[,1]
       }
     }
   } else {
@@ -242,7 +242,7 @@ make_xifti <- function(
     if (HCP_32k_auto_mwall & nrow(xifti$data$cortex_right) == 29716) {
       if (is.null(xifti$meta$cortex$medial_wall_mask$right)) {
         # from ciftiTools R/sysdata.rda
-        xifti$meta$cortex$medial_wall_mask$right <- HCP_32k_mwall_template[,2]
+        xifti$meta$cortex$medial_wall_mask$right <- ciftiTools.data$HCP_32k_mwall_template[,2]
       }
     }
   } else {